feat: extract component and some logic
